go.talks/present: separate parser from present tool
This will allow present files to be used by the tour and other tools.
This change is the first of a series. It focuses on moving the code
with a minimum of exported changes. I will refine the exported
interface of the present package in a followup change.

R=r, campoy
CC=golang-dev
http://codereview.appspot.com/6815075